+++ spinn_utilities/citation/__init__.py
@@ -1,6 +1,4 @@
 from .citation_updater_and_doi_generator import CitationUpdaterAndDoiGenerator
 from .citation_aggregator import CitationAggregator
-from .generate_aggregate import generate_aggregate
 
-__all__ = ["CitationUpdaterAndDoiGenerator", "CitationAggregator",
-           "generate_aggregate"]
+__all__ = ["CitationUpdaterAndDoiGenerator", "CitationAggregator"]
